refactor(query): Extract methods from runner/query.rs to explain, commits, version
Extract explain methods to runner/explain/mod.rs (~1,510 lines),
commits methods to runner/commits.rs (~890 lines),
version methods to runner/version.rs (~349 lines).

query.rs: 4,412 → 1,735 lines
