+++ crates/query/src/runner/query.rs
@@ -1,7 +1,6 @@
 //! Query execution methods for QueryRunner.
 
 use acp::{DocumentPermission, Identity};
-use document::Document;
 use identity::Did;
 use schema::CollectionVersion;
 use serde_json::{Map, Value as JsonValue};
@@ -12,12 +11,11 @@ use tracing::{debug, instrument, warn};
 use crate::document::{documents_to_plan_docs, documents_with_status_to_plan_docs};
 use crate::error::{QueryError, Result};
 use crate::mapper::{Requestable, Select};
-use crate::plan::PermissionFilterNode;
 use crate::planner::index_selection::{
     can_be_ordered_by_index, filter_to_index_scan, select_best_index,
 };
 use crate::planner::Planner;
-use crate::query_parse::{parse_query_with_variables, ExplainType};
+use crate::query_parse::parse_query_with_variables;
 use crate::txn::TransactionRegistry;
 
 use super::fetcher::FetcherWrapper;
@@ -55,1482 +53,6 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
             variables,
         )
         .await
-    }
-
-    /// Generate an explanation of the query plan.
-    ///
-    /// Used when queries include the @explain directive.
-    /// Supports three modes:
-    /// - Simple: Query plan structure without execution
-    /// - Execute: Run the query and return plan structure with execution metrics
-    /// - Debug: All plan nodes including internal ones
-    ///
-    /// Output format matches Go DefraDB:
-    /// ```json
-    /// {
-    ///   "explain": {
-    ///     "operationNode": [
-    ///       {
-    ///         "selectTopNode": {
-    ///           "selectNode": { ... "scanNode": { ... } }
-    ///         }
-    ///       }
-    ///     ]
-    ///   }
-    /// }
-    /// ```
-    pub async fn explain_query_with_identity(
-        &self,
-        query: &str,
-        caller_identity: Option<Did>,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        self.explain_query_with_identity_and_vars(query, caller_identity, explain_type, None)
-            .await
-    }
-
-    /// Generate an explanation of the query plan with variable support.
-    pub async fn explain_query_with_identity_and_vars(
-        &self,
-        query: &str,
-        caller_identity: Option<Did>,
-        explain_type: ExplainType,
-        variables: Option<&std::collections::HashMap<String, JsonValue>>,
-    ) -> Result<JsonValue> {
-        match explain_type {
-            ExplainType::Simple | ExplainType::Debug => {
-                // Simple and Debug modes: explain without execution
-                let selects = parse_query_with_variables(query, variables)?;
-                let mut operation_children: Vec<JsonValue> = Vec::new();
-
-                for select in selects {
-                    // Check if this is a top-level aggregate query (e.g., _avg, _count, _sum)
-                    let is_top_level_aggregate = Self::is_top_level_aggregate(&select);
-
-                    // Build the plan explanation for this select
-                    let select_node_content = self.explain_select(&select, explain_type).await?;
-
-                    if is_top_level_aggregate {
-                        // Top-level aggregates use topLevelNode wrapper
-                        let top_level_node = self.build_top_level_aggregate_explain(
-                            &select,
-                            select_node_content,
-                            explain_type,
-                        );
-                        operation_children.push(top_level_node);
-                    } else {
-                        // Regular queries use selectTopNode wrapper
-                        let select_top_node = serde_json::json!({
-                            "selectTopNode": select_node_content
-                        });
-                        operation_children.push(select_top_node);
-                    }
-                }
-
-                // Wrap all selects in operationNode array (Go's MultiNode pattern)
-                Ok(serde_json::json!({
-                    "explain": {
-                        "operationNode": operation_children
-                    }
-                }))
-            }
-            ExplainType::Execute => {
-                // Execute mode: run the query and collect metrics
-                self.execute_explain_with_vars(query, caller_identity, variables)
-                    .await
-            }
-        }
-    }
-
-    /// Generate an explanation of the mutation plan.
-    ///
-    /// Used when mutations include the @explain directive.
-    /// Output format matches Go DefraDB with createNode/deleteNode/updateNode/upsertNode.
-    pub async fn explain_mutation_with_identity(
-        &self,
-        mutation_str: &str,
-        caller_identity: Option<Did>,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        use crate::query_parse::parse_mutations;
-
-        match explain_type {
-            ExplainType::Simple | ExplainType::Debug => {
-                // Simple and Debug modes: explain without execution
-                let mutations = parse_mutations(mutation_str)?;
-                let mut operation_children: Vec<JsonValue> = Vec::new();
-
-                for mutation in mutations {
-                    let mutation_explain = self
-                        .explain_single_mutation(&mutation, explain_type)
-                        .await?;
-                    operation_children.push(mutation_explain);
-                }
-
-                // Wrap all mutations in operationNode array (Go's MultiNode pattern)
-                Ok(serde_json::json!({
-                    "explain": {
-                        "operationNode": operation_children
-                    }
-                }))
-            }
-            ExplainType::Execute => {
-                // Execute mode: run the mutation and collect metrics
-                self.execute_mutation_explain(mutation_str, caller_identity)
-                    .await
-            }
-        }
-    }
-
-    /// Execute mutation and return explain output with execution metrics.
-    async fn execute_mutation_explain(
-        &self,
-        mutation_str: &str,
-        caller_identity: Option<Did>,
-    ) -> Result<JsonValue> {
-        use crate::query_parse::parse_mutations;
-
-        let mutations = parse_mutations(mutation_str)?;
-        let mut operation_children: Vec<JsonValue> = Vec::new();
-        let mut execution_success = true;
-        let mut execution_errors: Vec<String> = Vec::new();
-
-        for mutation in mutations {
-            match self
-                .execute_single_mutation_with_metrics(&mutation, caller_identity.clone())
-                .await
-            {
-                Ok((mutation_explain, _doc_count, _exec_count)) => {
-                    operation_children.push(mutation_explain);
-                }
-                Err(e) => {
-                    execution_success = false;
-                    execution_errors.push(e.to_string());
-                }
-            }
-        }
-
-        // Go's executeAndExplainRequest calls Next() on the top-level operationNode.
-        // Each Next()=true yields one mutation result. After all mutations, Next()=false.
-        // So planExecutions = number_of_mutations + 1, sizeOfResult = number_of_mutations.
-        let num_mutations = operation_children.len() as u64;
-        let plan_executions = num_mutations + 1;
-        let size_of_result = num_mutations;
-
-        // Build explain result with execution metrics
-        let mut explain_result = Map::new();
-        explain_result.insert(
-            "operationNode".to_string(),
-            JsonValue::Array(operation_children),
-        );
-        explain_result.insert(
-            "executionSuccess".to_string(),
-            serde_json::json!(execution_success),
-        );
-        explain_result.insert(
-            "planExecutions".to_string(),
-            serde_json::json!(plan_executions),
-        );
-        explain_result.insert(
-            "sizeOfResult".to_string(),
-            serde_json::json!(size_of_result),
-        );
-
-        if !execution_errors.is_empty() {
-            explain_result.insert(
-                "executionErrors".to_string(),
-                serde_json::json!(execution_errors),
-            );
-        }
-
-        Ok(serde_json::json!({ "explain": JsonValue::Object(explain_result) }))
-    }
-
-    /// Execute a single mutation and return explain with metrics.
-    ///
-    /// Go's mutation plan runs as: mutationNode → selectTopNode → selectNode → scanNode.
-    /// The filter lives at the scanNode level (not selectNode). Metrics accumulate:
-    /// - Create: single pass AFTER creation
-    /// - Delete: single pass BEFORE deletion
-    /// - Update/Upsert: two passes (Phase 1: scan+mutate, Phase 2: scan+return), metrics sum
-    async fn execute_single_mutation_with_metrics(
-        &self,
-        mutation: &crate::mapper::Mutation,
-        caller_identity: Option<Did>,
-    ) -> Result<(JsonValue, usize, u64)> {
-        use crate::mapper::MutationType;
-
-        let node_kind = match mutation.mutation_type {
-            MutationType::Create => "createNode",
-            MutationType::Update => "updateNode",
-            MutationType::Delete => "deleteNode",
-            MutationType::Upsert => "upsertNode",
-        };
-
-        // Build a select with the mutation's filter/doc_ids for metric collection.
-        // Go puts the mutation filter on the scanNode; build_plan places the filter
-        // on both scanNode (for iteration/docFetch counting) and selectNode (for
-        // filterMatches counting), which matches Go's behavior.
-        let mut metric_select = Select::new(&mutation.collection_name);
-        if let Some(ref filter) = mutation.filter {
-            metric_select.filter = Some(filter.clone());
-        }
-        if let Some(ref doc_ids) = mutation.doc_ids {
-            metric_select.doc_ids = Some(doc_ids.clone());
-        }
-
-        // Phase 1: Collect metrics BEFORE mutation (delete, update, upsert).
-        // For delete: this is the only scan (single pass over original data).
-        // For update/upsert: Phase 1 captures the "find + mutate" scan metrics.
-        let phase1 = if mutation.mutation_type != MutationType::Create {
-            Some(
-                self.execute_select_with_metrics(&metric_select, caller_identity.clone())
-                    .await?,
-            )
-        } else {
-            None
-        };
-
-        // Execute the actual mutation
-        if let Some(ref mutator) = self.mutator {
-            use chrono::{FixedOffset, Utc};
-            let utc_offset = FixedOffset::east_opt(0).unwrap();
-            let request_time = Utc::now().with_timezone(&utc_offset);
-
-            let collection = self.get_collection(&mutation.collection_name).await?;
-            let mapping = self.build_mutation_mapping(mutation)?;
-            let resolved_doc_ids = self.resolve_filter_to_doc_ids(mutation).await?;
-
-            let mut plan: Box<dyn crate::planner::PlanNode> = match mutation.mutation_type {
-                MutationType::Create => {
-                    let inputs = self.build_create_inputs(mutation)?;
-                    Box::new(
-                        crate::plan::CreateNode::new(
-                            &mutation.collection_name,
-                            mutator.clone(),
-                            mapping.clone(),
-                        )
-                        .with_collection(collection.clone())
-                        .with_request_time(request_time)
-                        .with_inputs(inputs),
-                    )
-                }
-                MutationType::Update => {
-                    let input = self.build_update_input(mutation)?;
-                    let fetcher: Arc<dyn crate::fetcher::DocFetcher> = self.fetcher.clone();
-                    let mut node = crate::plan::UpdateNode::new(
-                        &mutation.collection_name,
-                        mutator.clone(),
-                        fetcher,
-                        mapping.clone(),
-                    )
-                    .with_collection(collection.clone())
-                    .with_request_time(request_time)
-                    .with_input(input);
-
-                    if let Some(ref doc_ids) = resolved_doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    } else if let Some(ref doc_ids) = mutation.doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    }
-                    if let Some(ref filter) = mutation.filter {
-                        node = node.with_filter(filter.clone());
-                    }
-                    Box::new(node)
-                }
-                MutationType::Delete => {
-                    let fetcher: Arc<dyn crate::fetcher::DocFetcher> = self.fetcher.clone();
-                    let mut node = crate::plan::DeleteNode::new(
-                        &mutation.collection_name,
-                        mutator.clone(),
-                        fetcher,
-                        mapping.clone(),
-                    );
-
-                    if let Some(ref doc_ids) = resolved_doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    } else if let Some(ref doc_ids) = mutation.doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    }
-                    if mutation.filter.is_some()
-                        && resolved_doc_ids.is_none()
-                        && mutation.doc_ids.is_none()
-                    {
-                        node = node.with_filter(mutation.filter.clone().unwrap());
-                    }
-                    Box::new(node)
-                }
-                MutationType::Upsert => {
-                    let mut node = crate::plan::UpsertNode::new(
-                        &mutation.collection_name,
-                        mutator.clone(),
-                        mapping.clone(),
-                    )
-                    .with_collection(collection.clone())
-                    .with_request_time(request_time);
-
-                    if !mutation.create_input.is_empty() {
-                        let create_input =
-                            self.build_upsert_input_from_map(&mutation.create_input[0])?;
-                        node = node.with_create_input(create_input);
-                    }
-                    if !mutation.update_input.is_empty() {
-                        let update_input =
-                            self.build_upsert_input_from_map(&mutation.update_input)?;
-                        node = node.with_update_input(update_input);
-                    }
-                    if let Some(ref doc_ids) = resolved_doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    } else if let Some(ref doc_ids) = mutation.doc_ids {
-                        node = node.with_doc_ids(doc_ids.clone());
-                    }
-                    Box::new(node)
-                }
-            };
-
-            // Execute the mutation plan (ignore results, we just need the side effects)
-            plan.init().await?;
-            plan.start().await?;
-            while plan.next().await? {}
-            plan.close().await?;
-        }
-
-        // Phase 2: Collect metrics AFTER mutation (create, update, upsert).
-        // For create: this is the only scan (single pass over created data).
-        // For update/upsert: Phase 2 captures the "return results" scan metrics.
-        let phase2 = if mutation.mutation_type != MutationType::Delete {
-            Some(
-                self.execute_select_with_metrics(&metric_select, caller_identity.clone())
-                    .await?,
-            )
-        } else {
-            None
-        };
-
-        // Combine metrics from phases
-        let combined_explain = match (&phase1, &phase2) {
-            (Some((p1, _, _)), Some((p2, _, _))) => {
-                // Two-pass (update/upsert): merge by summing all numeric values
-                Self::merge_execute_metrics(p1, p2)
-            }
-            (Some((p1, _, _)), None) => p1.clone(),
-            (None, Some((p2, _, _))) => p2.clone(),
-            _ => unreachable!(),
-        };
-
-        // Wrap in selectTopNode
-        let select_node_content = Self::ensure_select_node_wrapper(
-            combined_explain,
-            &metric_select,
-            ExplainType::Execute,
-        );
-
-        // Build mutation node
-        let mut mutation_inner = serde_json::Map::new();
-
-        // Mutation-specific fields
-        match mutation.mutation_type {
-            MutationType::Create => {
-                let (_, _, plan_execs) = phase2.as_ref().unwrap();
-                mutation_inner.insert("iterations".to_string(), serde_json::json!(*plan_execs));
-            }
-            MutationType::Delete => {
-                let (_, _, plan_execs) = phase1.as_ref().unwrap();
-                mutation_inner.insert("iterations".to_string(), serde_json::json!(*plan_execs));
-            }
-            MutationType::Update => {
-                let (_, result_count, plan_execs) = phase1.as_ref().unwrap();
-                mutation_inner.insert("iterations".to_string(), serde_json::json!(*plan_execs));
-                mutation_inner.insert(
-                    "updates".to_string(),
-                    serde_json::json!(*result_count as u64),
-                );
-            }
-            MutationType::Upsert => {
-                // Go's upsertNode returns empty map for execute explain (no iterations)
-            }
-        }
-
-        mutation_inner.insert("selectTopNode".to_string(), select_node_content);
-
-        let mutation_node = serde_json::json!({
-            node_kind: mutation_inner
-        });
-
-        let doc_count = match (&phase1, &phase2) {
-            (_, Some((_, count, _))) => *count,
-            (Some((_, count, _)), None) => *count,
-            _ => 0,
-        };
-
-        Ok((mutation_node, doc_count, 1))
-    }
-
-    /// Recursively merge two execute explain JSON trees by summing numeric values.
-    /// Used to combine Phase 1 and Phase 2 metrics for update/upsert mutations.
-    fn merge_execute_metrics(phase1: &JsonValue, phase2: &JsonValue) -> JsonValue {
-        match (phase1, phase2) {
-            (JsonValue::Object(a), JsonValue::Object(b)) => {
-                let mut merged = serde_json::Map::new();
-                for (key, val_a) in a {
-                    if let Some(val_b) = b.get(key) {
-                        merged.insert(key.clone(), Self::merge_execute_metrics(val_a, val_b));
-                    } else {
-                        merged.insert(key.clone(), val_a.clone());
-                    }
-                }
-                for (key, val_b) in b {
-                    if !a.contains_key(key) {
-                        merged.insert(key.clone(), val_b.clone());
-                    }
-                }
-                JsonValue::Object(merged)
-            }
-            (JsonValue::Number(a), JsonValue::Number(b)) => {
-                let sum = a.as_u64().unwrap_or(0) + b.as_u64().unwrap_or(0);
-                serde_json::json!(sum)
-            }
-            _ => phase2.clone(),
-        }
-    }
-
-    /// Generate an explanation for a single mutation operation.
-    async fn explain_single_mutation(
-        &self,
-        mutation: &crate::mapper::Mutation,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        use crate::mapper::MutationType;
-
-        // Get the mutation node kind name
-        let node_kind = match mutation.mutation_type {
-            MutationType::Create => "createNode",
-            MutationType::Update => "updateNode",
-            MutationType::Delete => "deleteNode",
-            MutationType::Upsert => "upsertNode",
-        };
-
-        // Build the inner select plan explanation
-        // Mutations in Go have: mutationNode -> selectTopNode -> selectNode -> scanNode
-        let collection = self
-            .collection_provider
-            .get_collection(&mutation.collection_name)
-            .await?
-            .ok_or_else(|| QueryError::collection_not_found(&mutation.collection_name))?;
-
-        // Build a select for the mutation's result fields, including filter and docIDs
-        // These are passed through to the scanNode for proper explain output
-        let mut select = crate::mapper::Select::new(&mutation.collection_name);
-        if let Some(ref filter) = mutation.filter {
-            select = select.with_filter(filter.clone());
-        }
-        if let Some(ref doc_ids) = mutation.doc_ids {
-            select = select.with_doc_ids(doc_ids.clone());
-        }
-        let inner_explain = self.explain_simple_select(&select, &collection, explain_type)?;
-
-        // Build mutation-specific attributes
-        let mut mutation_attrs = serde_json::Map::new();
-
-        match mutation.mutation_type {
-            MutationType::Create => {
-                // input: array of input objects
-                let input_array: Vec<JsonValue> = mutation
-                    .create_input
-                    .iter()
-                    .map(|input| {
-                        let mut input_obj = serde_json::Map::new();
-                        for (field_name, value) in input {
-                            input_obj.insert(field_name.clone(), value.clone());
-                        }
-                        JsonValue::Object(input_obj)
-                    })
-                    .collect();
-                mutation_attrs.insert("input".to_string(), JsonValue::Array(input_array));
-            }
-            MutationType::Update => {
-                // docID: array of doc IDs (or null)
-                if let Some(ref doc_ids) = mutation.doc_ids {
-                    mutation_attrs.insert(
-                        "docID".to_string(),
-                        JsonValue::Array(
-                            doc_ids
-                                .iter()
-                                .map(|id| JsonValue::String(id.clone()))
-                                .collect(),
-                        ),
-                    );
-                } else {
-                    mutation_attrs.insert("docID".to_string(), JsonValue::Null);
-                }
-
-                // filter: filter expression (or null, including empty filter)
-                if let Some(ref filter) = mutation.filter {
-                    let conditions = filter.conditions();
-                    if conditions.is_empty() {
-                        mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                    } else {
-                        mutation_attrs.insert("filter".to_string(), serde_json::json!(conditions));
-                    }
-                } else {
-                    mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                }
-
-                // input: map of update values
-                let mut input_obj = serde_json::Map::new();
-                for (field_name, value) in &mutation.update_input {
-                    input_obj.insert(field_name.clone(), value.clone());
-                }
-                mutation_attrs.insert("input".to_string(), JsonValue::Object(input_obj));
-            }
-            MutationType::Delete => {
-                // docID: array of doc IDs (or null)
-                if let Some(ref doc_ids) = mutation.doc_ids {
-                    mutation_attrs.insert(
-                        "docID".to_string(),
-                        JsonValue::Array(
-                            doc_ids
-                                .iter()
-                                .map(|id| JsonValue::String(id.clone()))
-                                .collect(),
-                        ),
-                    );
-                } else {
-                    mutation_attrs.insert("docID".to_string(), JsonValue::Null);
-                }
-
-                // filter: filter expression (or null, including empty filter)
-                if let Some(ref filter) = mutation.filter {
-                    let conditions = filter.conditions();
-                    if conditions.is_empty() {
-                        mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                    } else {
-                        mutation_attrs.insert("filter".to_string(), serde_json::json!(conditions));
-                    }
-                } else {
-                    mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                }
-            }
-            MutationType::Upsert => {
-                // Go format: separate create, filter, and update fields
-                // create: map of fields for new document creation
-                if !mutation.create_input.is_empty() {
-                    let mut create_obj = serde_json::Map::new();
-                    for (field_name, value) in &mutation.create_input[0] {
-                        create_obj.insert(field_name.clone(), value.clone());
-                    }
-                    mutation_attrs.insert("create".to_string(), JsonValue::Object(create_obj));
-                }
-
-                // filter: filter expression used to find existing documents
-                if let Some(ref filter) = mutation.filter {
-                    let conditions = filter.conditions();
-                    if conditions.is_empty() {
-                        mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                    } else {
-                        mutation_attrs.insert("filter".to_string(), serde_json::json!(conditions));
-                    }
-                } else {
-                    mutation_attrs.insert("filter".to_string(), JsonValue::Null);
-                }
-
-                // update: map of fields for updating existing document
-                if !mutation.update_input.is_empty() {
-                    let mut update_obj = serde_json::Map::new();
-                    for (field_name, value) in &mutation.update_input {
-                        update_obj.insert(field_name.clone(), value.clone());
-                    }
-                    mutation_attrs.insert("update".to_string(), JsonValue::Object(update_obj));
-                }
-            }
-        }
-
-        // Add selectTopNode containing the inner select explanation
-        mutation_attrs.insert("selectTopNode".to_string(), inner_explain);
-
-        // Wrap in the mutation node type
-        let mutation_node = serde_json::json!({
-            node_kind: JsonValue::Object(mutation_attrs)
-        });
-
-        Ok(mutation_node)
-    }
-
-    /// Execute the query with variables and return explain output with execution metrics.
-    /// Format matches Go DefraDB's executeAndExplainRequest output.
-    async fn execute_explain_with_vars(
-        &self,
-        query: &str,
-        caller_identity: Option<Did>,
-        variables: Option<&std::collections::HashMap<String, JsonValue>>,
-    ) -> Result<JsonValue> {
-        let selects = parse_query_with_variables(query, variables)?;
-
-        let mut operation_children: Vec<JsonValue> = Vec::new();
-        let mut execution_success = true;
-        let mut execution_errors: Vec<String> = Vec::new();
-
-        for select in selects {
-            let is_top_level_aggregate = Self::is_top_level_aggregate(&select);
-
-            // Execute the select and collect metrics
-            match self
-                .execute_select_with_metrics(&select, caller_identity.clone())
-                .await
-            {
-                Ok((explanation, _doc_count, _exec_count)) => {
-                    // Ensure selectNode wrapper
-                    let select_node_content = Self::ensure_select_node_wrapper(
-                        explanation,
-                        &select,
-                        ExplainType::Execute,
-                    );
-
-                    if is_top_level_aggregate {
-                        // Top-level aggregates use topLevelNode wrapper
-                        let top_level_node = self.build_top_level_aggregate_explain(
-                            &select,
-                            select_node_content,
-                            ExplainType::Execute,
-                        );
-                        operation_children.push(top_level_node);
-                    } else {
-                        // Regular queries use selectTopNode wrapper
-                        let select_top_node = serde_json::json!({
-                            "selectTopNode": select_node_content
-                        });
-                        operation_children.push(select_top_node);
-                    }
-                }
-                Err(e) => {
-                    execution_success = false;
-                    execution_errors.push(e.to_string());
-                }
-            }
-        }
-
-        // Go's executeAndExplainRequest calls Next() on the top-level operationNode.
-        // Each Next()=true yields one query result. After all queries, Next()=false.
-        // So planExecutions = number_of_queries + 1, sizeOfResult = number_of_queries.
-        let num_queries = operation_children.len() as u64;
-        let plan_executions = num_queries + 1;
-        let size_of_result = num_queries;
-
-        // Build explain result with operationNode and execution metrics
-        let mut explain_result = Map::new();
-        explain_result.insert(
-            "operationNode".to_string(),
-            JsonValue::Array(operation_children),
-        );
-        explain_result.insert(
-            "executionSuccess".to_string(),
-            serde_json::json!(execution_success),
-        );
-        explain_result.insert(
-            "planExecutions".to_string(),
-            serde_json::json!(plan_executions),
-        );
-        explain_result.insert(
-            "sizeOfResult".to_string(),
-            serde_json::json!(size_of_result),
-        );
-
-        if !execution_errors.is_empty() {
-            explain_result.insert(
-                "executionErrors".to_string(),
-                serde_json::json!(execution_errors),
-            );
-        }
-
-        Ok(serde_json::json!({ "explain": JsonValue::Object(explain_result) }))
-    }
-
-    /// Execute a select and return the explain output with metrics.
-    /// Format matches Go DefraDB's collectExecuteExplainInfo output.
-    async fn execute_select_with_metrics(
-        &self,
-        select: &Select,
-        caller_identity: Option<Did>,
-    ) -> Result<(JsonValue, usize, u64)> {
-        // Handle _commits system collection (no real collection exists)
-        if select.collection_name == "_commits" {
-            // Actually execute the commits query to get real metrics
-            let results = self.execute_commits_query(select).await?;
-            let doc_count = results.as_array().map(|a| a.len()).unwrap_or(0);
-
-            // Build execute explain with real metrics matching Go's format:
-            // dagScanNode.iterations = doc_count + 1 (includes terminal Next()=false)
-            // selectNode.iterations = doc_count + 1
-            // selectNode.filterMatches = doc_count
-            let iterations = (doc_count as u64) + 1;
-            let explanation = serde_json::json!({
-                "selectNode": {
-                    "filterMatches": doc_count as u64,
-                    "iterations": iterations,
-                    "dagScanNode": {
-                        "iterations": iterations,
-                    }
-                }
-            });
-            return Ok((explanation, doc_count, 1));
-        }
-
-        // Get collection schema
-        let collection = self
-            .collection_provider
-            .get_collection(&select.collection_name)
-            .await?
-            .ok_or_else(|| QueryError::collection_not_found(&select.collection_name))?;
-
-        // Embedded-only types (interface types from view SDL) are not root-queryable
-        if collection.is_embedded_only {
-            return Err(QueryError::parse(format!(
-                "Cannot query field \"{}\" on type \"Query\".",
-                select.collection_name
-            )));
-        }
-
-        let fetcher = self.fetcher.as_ref();
-
-        // Check if we can use an index (filter-based or ordering-based)
-        let can_use_filter_index = select.doc_ids.is_none()
-            && select.filter.is_some()
-            && !collection.indexes.is_empty()
-            && fetcher.supports_index_queries()
-            && select
-                .filter
-                .as_ref()
-                .map(|f| select_best_index(f, &collection.indexes).is_some())
-                .unwrap_or(false);
-
-        // Also use index when it can provide ordering (even without a filter)
-        let can_use_ordering_index = select.doc_ids.is_none()
-            && select.order_by.is_some()
-            && !collection.indexes.is_empty()
-            && fetcher.supports_index_queries()
-            && select
-                .order_by
-                .as_ref()
-                .map(|o| {
-                    collection
-                        .indexes
-                        .iter()
-                        .any(|idx| can_be_ordered_by_index(o, idx).0)
-                })
-                .unwrap_or(false);
-
-        let can_use_index = can_use_filter_index || can_use_ordering_index;
-
-        // Check if any aggregates reference relations (e.g., _sum(articles: {field: pages}))
-        // Relation aggregates need the Planner to create TypeJoinMany nodes
-        let has_relation_aggregates = select.fields.iter().any(|f| {
-            if let Requestable::Aggregate(agg) = f {
-                agg.targets
-                    .iter()
-                    .any(|t| !t.host_name.is_empty() && t.host_name != select.collection_name)
-            } else {
-                false
-            }
-        });
-
-        // Check if this query has nested selections (relations, _group, etc.)
-        // These require the Planner to construct proper join/group nodes.
-        let has_nested = select
-            .fields
-            .iter()
-            .any(|f| matches!(f, Requestable::Select(_)));
-
-        let filter_has_relations = select
-            .filter
-            .as_ref()
-            .map(|f| f.has_relation_filters())
-            .unwrap_or(false);
-
-        let order_has_relations = select
-            .order_by
-            .as_ref()
-            .map(|o| o.has_relation_order())
-            .unwrap_or(false);
-
-        let has_similarity = select
-            .fields
-            .iter()
-            .any(|f| matches!(f, Requestable::Similarity(_)));
-
-        if can_use_index
-            || has_relation_aggregates
-            || has_nested
-            || filter_has_relations
-            || has_similarity
-        {
-            // Use Planner path for index-based queries, relation aggregates,
-            // relation filters/ordering, or similarity
-            let fetcher_arc = FetcherWrapper::new(fetcher);
-            let collections_map = self.collections_map().await?;
-            let collections: Vec<CollectionVersion> =
-                collections_map.values().map(|c| (**c).clone()).collect();
-
-            let mut planner = Planner::new(collections).with_fetcher(Arc::new(fetcher_arc));
-            if let Some(ref lens_store) = self.lens_store {
-                planner = planner.with_lens_store(lens_store.clone());
-            }
-            let plan_result = planner.plan_with_index_info(select)?;
-            let mut plan = plan_result.plan;
-
-            // Wrap with permission filter if needed (explain path)
-            if let (Some(ref acp), Some(ref policy)) = (&self.acp, &collection.policy) {
-                plan = Box::new(PermissionFilterNode::new(
-                    plan,
-                    acp.clone(),
-                    Identity::from(caller_identity),
-                    &policy.id,
-                    &policy.resource_name,
-                ));
-            }
-
-            // Execute the plan and count iterations
-            plan.init().await?;
-            plan.start().await?;
-
-            // Go counts ALL next() calls (including the final false) for planExecutions
-            let mut plan_executions: u64 = 0;
-            let mut result_count: usize = 0;
-
-            loop {
-                plan_executions += 1;
-                if !plan.next().await? {
-                    break;
-                }
-                result_count += 1;
-            }
-
-            plan.close().await?;
-
-            // Use explain_execute to get metrics from each node
-            let explanation = plan.explain_execute();
-
-            Ok((explanation, result_count, plan_executions))
-        } else {
-            // Standard path: fetch all docs and build scan-based plan
-            let mapping = plan::build_mapping(select, &collection)?;
-
-            // When show_deleted is true, we need to use get_all_with_deleted to include
-            // logically deleted documents. The doc_ids filter will be applied by the SelectNode.
-            let plan_docs = if select.show_deleted {
-                let docs_with_status = fetcher
-                    .get_all_with_deleted(&select.collection_name, true)
-                    .await?;
-                documents_with_status_to_plan_docs(&docs_with_status, &mapping)?
-            } else if let Some(ref doc_ids) = select.doc_ids {
-                // Deduplicate doc_ids while preserving order (Go compatibility)
-                let mut seen = HashSet::new();
-                let unique_ids: Vec<String> = doc_ids
-                    .iter()
-                    .filter(|id| seen.insert((*id).clone()))
-                    .cloned()
-                    .collect();
-                let result = fetcher
-                    .get_by_ids(&select.collection_name, &unique_ids)
-                    .await?;
-                documents_to_plan_docs(&result.into_docs(), &mapping)?
-            } else {
-                let docs = fetcher.get_all(&select.collection_name).await?;
-                documents_to_plan_docs(&docs, &mapping)?
-            };
-            let doc_count = plan_docs.len();
-
-            // Build ACP filter config if collection has policy and ACP is configured
-            let acp_filter =
-                if let (Some(ref acp), Some(ref policy)) = (&self.acp, &collection.policy) {
-                    Some(plan::AcpFilter {
-                        acp: acp.clone(),
-                        identity: Identity::from(caller_identity),
-                        policy_id: policy.id.clone(),
-                        resource_name: policy.resource_name.clone(),
-                    })
-                } else {
-                    None
-                };
-
-            // Build the plan (ACP filter is inserted inside, after Select but before aggregates)
-            let mut plan = plan::build_plan(
-                select,
-                plan_docs.clone(),
-                mapping.clone(),
-                &collection,
-                acp_filter,
-            )?;
-
-            // Execute the plan and count iterations
-            plan.init().await?;
-            plan.start().await?;
-
-            // Go counts ALL next() calls (including the final false) for planExecutions
-            let mut plan_executions: u64 = 0;
-            let mut result_count: usize = 0;
-
-            loop {
-                plan_executions += 1;
-                if !plan.next().await? {
-                    break;
-                }
-                result_count += 1;
-            }
-
-            plan.close().await?;
-
-            // Use explain_execute to get metrics from each node
-            let explanation = plan.explain_execute();
-
-            Ok((explanation, result_count, plan_executions))
-        }
-    }
-
-    /// Add execution metrics to the explain output (Go format).
-    /// Metrics are added to the scanNode, not the selectNode.
-    fn add_iterations_to_explain(
-        mut explanation: JsonValue,
-        iterations: u64,
-        doc_fetches: usize,
-        field_count: usize,
-    ) -> JsonValue {
-        // The explanation is { "selectNode": { "scanNode": { ... } } }
-        // We need to add metrics to the scanNode
-        Self::add_metrics_to_scan_node(&mut explanation, iterations, doc_fetches, field_count);
-        explanation
-    }
-
-    /// Recursively find and add metrics to scanNode.
-    /// If the scanNode has an indexName field, it's an index scan and gets indexFetches.
-    fn add_metrics_to_scan_node(
-        value: &mut JsonValue,
-        iterations: u64,
-        doc_fetches: usize,
-        field_count: usize,
-    ) {
-        if let Some(obj) = value.as_object_mut() {
-            // Check if this object contains scanNode
-            if let Some(scan_node) = obj.get_mut("scanNode") {
-                if let Some(scan_obj) = scan_node.as_object_mut() {
-                    scan_obj.insert(
-                        "iterations".to_string(),
-                        serde_json::json!(iterations as u64),
-                    );
-                    scan_obj.insert(
-                        "docFetches".to_string(),
-                        serde_json::json!(doc_fetches as u64),
-                    );
-                    // fieldFetches = number of fields per doc * number of docs fetched
-                    let field_fetches = (field_count * doc_fetches) as u64;
-                    scan_obj.insert("fieldFetches".to_string(), serde_json::json!(field_fetches));
-
-                    // indexFetches is set by IndexScanNode::explain_inner() with
-                    // the actual index key lookup count. For regular scans without an
-                    // index, default to 0 (Go always includes this property).
-                    if !scan_obj.contains_key("indexFetches") {
-                        scan_obj.insert("indexFetches".to_string(), serde_json::json!(0u64));
-                    }
-                    return;
-                }
-            }
-
-            // Recurse into child objects
-            for (_, child) in obj.iter_mut() {
-                Self::add_metrics_to_scan_node(child, iterations, doc_fetches, field_count);
-            }
-        }
-    }
-
-    /// Generate an explanation of a single Select operation.
-    async fn explain_select(
-        &self,
-        select: &Select,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        // Handle encrypted search queries - return a simple seScanNode explanation
-        if select.is_encrypted {
-            return Ok(serde_json::json!({
-                "selectNode": {
-                    "seScanNode": {
-                        "collection": select.collection_name,
-                        "filter": select.filter.as_ref().map(|f| f.conditions())
-                    }
-                }
-            }));
-        }
-
-        // Handle _commits system collection specially
-        if select.collection_name == "_commits" {
-            return self.explain_commits_select(select, explain_type);
-        }
-
-        // Get collection schema
-        let collection = self
-            .collection_provider
-            .get_collection(&select.collection_name)
-            .await?
-            .ok_or_else(|| QueryError::collection_not_found(&select.collection_name))?;
-
-        // Check if this query has nested selections (relations)
-        let has_nested = select
-            .fields
-            .iter()
-            .any(|f| matches!(f, Requestable::Select(_)));
-
-        // Check if an ordering-only index can be used (planner needed for IndexScanNode)
-        let has_ordering_index = !has_nested
-            && select.order_by.is_some()
-            && !collection.indexes.is_empty()
-            && select
-                .order_by
-                .as_ref()
-                .map(|o| {
-                    collection
-                        .indexes
-                        .iter()
-                        .any(|idx| can_be_ordered_by_index(o, idx).0)
-                })
-                .unwrap_or(false);
-
-        // Check if a filter-based index can be used
-        let has_filter_index = !has_nested
-            && select.filter.is_some()
-            && !collection.indexes.is_empty()
-            && select
-                .filter
-                .as_ref()
-                .map(|f| select_best_index(f, &collection.indexes).is_some())
-                .unwrap_or(false);
-
-        // Check if any aggregates reference relations (e.g., _sum(articles: {field: pages}))
-        let has_relation_aggregates = select.fields.iter().any(|f| {
-            if let Requestable::Aggregate(agg) = f {
-                agg.targets
-                    .iter()
-                    .any(|t| !t.host_name.is_empty() && t.host_name != select.collection_name)
-            } else {
-                false
-            }
-        });
-
-        // Check if the filter references relation fields (e.g., {author: {verified: true}})
-        let filter_has_relations = select
-            .filter
-            .as_ref()
-            .map(|f| f.has_relation_filters())
-            .unwrap_or(false);
-
-        // Check if the order references relation fields (e.g., {author: {age: DESC}})
-        let order_has_relations = select
-            .order_by
-            .as_ref()
-            .map(|o| o.has_relation_order())
-            .unwrap_or(false);
-
-        // Check if any similarity fields are present (require SimilarityNode in planner)
-        let has_similarity = select
-            .fields
-            .iter()
-            .any(|f| matches!(f, Requestable::Similarity(_)));
-
-        // Check if any secondary relation ID fields are selected (e.g., `_authorID`)
-        let has_secondary_relation_id = select.fields.iter().any(|f| {
-            if let Requestable::Field(field) = f {
-                let field_name = &field.name;
-                if field_name.starts_with('_') && field_name.ends_with("ID") && field_name.len() > 3
-                {
-                    let relation_name = &field_name[1..field_name.len() - 2];
-                    if let Some(relation_field) = collection.field_by_name(relation_name) {
-                        return relation_field.kind.is_relation() && !relation_field.is_primary;
-                    }
-                }
-            }
-            false
-        });
-
-        let is_view = collection.query.is_some();
-
-        if is_view
-            || has_nested
-            || has_ordering_index
-            || has_filter_index
-            || has_relation_aggregates
-            || filter_has_relations
-            || order_has_relations
-            || has_similarity
-            || has_secondary_relation_id
-        {
-            // Use the Planner for views, nested selections, index usage, relation aggregates,
-            // relation filters/ordering, similarity, or secondary relation IDs
-            self.explain_nested_select(select, explain_type).await
-        } else {
-            // Explain simple query plan
-            self.explain_simple_select(select, &collection, explain_type)
-        }
-    }
-
-    /// Generate an explanation for a query with nested selections.
-    async fn explain_nested_select(
-        &self,
-        select: &Select,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        // Build the plan using the Planner
-        let collection_names = self.collection_provider.list_collections().await?;
-        let mut collections = Vec::new();
-        for name in collection_names {
-            if let Some(coll) = self.collection_provider.get_collection(&name).await? {
-                collections.push((*coll).clone());
-            }
-        }
-
-        let mut planner = Planner::new(collections);
-        if let Some(ref lens_store) = self.lens_store {
-            planner = planner.with_lens_store(lens_store.clone());
-        }
-        let plan_result = planner.plan_with_index_info(select)?;
-        let plan = plan_result.plan;
-
-        // Get the plan explanation based on type
-        let explain = match explain_type {
-            ExplainType::Debug => plan.explain_debug(),
-            _ => plan.explain(),
-        };
-
-        // Ensure result is wrapped in selectNode (Go format)
-        Ok(Self::ensure_select_node_wrapper(
-            explain,
-            select,
-            explain_type,
-        ))
-    }
-
-    /// Generate an explanation for a simple query without nested selections.
-    fn explain_simple_select(
-        &self,
-        select: &Select,
-        collection: &CollectionVersion,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        // Build document mapping and plan
-        let mapping = plan::build_mapping(select, collection)?;
-
-        // Create an empty plan with no documents for explanation purposes
-        let plan = plan::build_plan(select, vec![], mapping, collection, None)?;
-
-        // Get the plan explanation based on type
-        let explain = match explain_type {
-            ExplainType::Debug => plan.explain_debug(),
-            _ => plan.explain(),
-        };
-
-        // Ensure result is wrapped in selectNode (Go format)
-        Ok(Self::ensure_select_node_wrapper(
-            explain,
-            select,
-            explain_type,
-        ))
-    }
-
-    /// Process explain output for Go format compatibility.
-    ///
-    /// Since we now always create SelectNode in the plan, this function handles:
-    /// - For Simple mode: ensures docID and filter attributes are in selectNode
-    /// - For Debug mode: returns as-is (no additional attributes)
-    /// - For Execute mode: returns as-is (attributes added during execution)
-    fn ensure_select_node_wrapper(
-        explain: JsonValue,
-        _select: &Select,
-        explain_type: ExplainType,
-    ) -> JsonValue {
-        // For Debug mode, return as-is (Go debug doesn't add attributes)
-        if matches!(explain_type, ExplainType::Debug) {
-            return explain;
-        }
-
-        // For Simple/Execute mode, the SelectNode already has the attributes
-        // from its explain_inner method, so return as-is
-        explain
-    }
-
-    /// Generate an explanation for a _commits system collection query.
-    ///
-    /// Returns a dagScanNode structure matching Go's explain output for commits queries.
-    fn explain_commits_select(
-        &self,
-        select: &Select,
-        explain_type: ExplainType,
-    ) -> Result<JsonValue> {
-        // For Debug mode, return empty inner objects
-        if matches!(explain_type, ExplainType::Debug) {
-            return Ok(serde_json::json!({
-                "selectNode": {
-                    "dagScanNode": {}
-                }
-            }));
-        }
-
-        // Build the dagScanNode attributes for Simple/Execute mode
-        let mut dag_scan_attrs = serde_json::Map::new();
-
-        // cid: the specific commit CID if provided, else null
-        if let Some(ref cid) = select.cid {
-            dag_scan_attrs.insert("cid".to_string(), serde_json::json!(cid));
-        } else {
-            dag_scan_attrs.insert("cid".to_string(), serde_json::Value::Null);
-        }
-
-        // prefixes: array of storage prefixes being scanned
-        // Format: /d/<docID> for document-specific commits
-        let prefixes: Vec<String> = if let Some(ref doc_ids) = select.doc_ids {
-            doc_ids.iter().map(|id| format!("/d/{}", id)).collect()
-        } else {
-            vec![]
-        };
-        dag_scan_attrs.insert("prefixes".to_string(), serde_json::json!(prefixes));
-
-        // Build the selectNode wrapper (Go structure: selectNode -> dagScanNode)
-        let dag_scan_node = serde_json::json!({ "dagScanNode": dag_scan_attrs });
-
-        Ok(serde_json::json!({
-            "selectNode": dag_scan_node
-        }))
-    }
-
-    /// Check if a select represents a top-level aggregate query (e.g., _avg, _count, _sum).
-    ///
-    /// Top-level aggregates are queries like `_count(Author)` where the aggregate function
-    /// is the root query field name itself.
-    ///
-    /// This does NOT include queries like `Author { _count(books) }` - those are regular
-    /// queries with aggregate sub-fields, not top-level aggregates.
-    fn is_top_level_aggregate(select: &Select) -> bool {
-        // Only true when the query field name itself is an aggregate function
-        let field_name = select.field.name.as_str();
-        field_name.starts_with('_')
-            && ["_count", "_sum", "_avg", "_min", "_max"].contains(&field_name)
-    }
-
-    /// Aggregate node kind names that can wrap a selectNode in the plan explain.
-    const AGGREGATE_NODE_KINDS: &'static [&'static str] =
-        &["countNode", "sumNode", "averageNode", "minNode", "maxNode"];
-
-    /// Aggregate-specific explain fields that should be stripped when unwrapping aggregate nodes.
-    /// "sources" appears in default explain, "iterations" in execute explain.
-    const AGGREGATE_EXPLAIN_FIELDS: [&'static str; 2] = ["sources", "iterations"];
-
-    /// Strip aggregate wrapper nodes from explain output for top-level aggregate queries.
-    ///
-    /// The Rust planner wraps the plan in aggregate nodes (e.g., CountNode → SelectNode → ScanNode),
-    /// but Go's explain format puts aggregates as siblings in topLevelNode, not as wrappers.
-    /// This function peels off any top-level aggregate wrappers to expose the inner selectNode.
-    ///
-    /// Example: `{ "countNode": { "sources": [...], "selectNode": { "scanNode": {...} } } }`
-    /// becomes: `{ "selectNode": { "scanNode": {...} } }`
-    fn strip_aggregate_wrappers(mut explain: JsonValue) -> JsonValue {
-        loop {
-            // Check if this is an aggregate wrapper node
-            let is_aggregate_wrapper = if let Some(obj) = explain.as_object() {
-                // An aggregate wrapper has the aggregate node kind as the only top-level key
-                obj.len() == 1
-                    && obj
-                        .keys()
-                        .next()
-                        .map(|k| Self::AGGREGATE_NODE_KINDS.contains(&k.as_str()))
-                        .unwrap_or(false)
-            } else {
-                false
-            };
-
-            if is_aggregate_wrapper {
-                // Unwrap: take the inner value from the aggregate node
-                let obj = explain.as_object_mut().unwrap();
-                let key = obj.keys().next().unwrap().clone();
-                explain = obj.remove(&key).unwrap();
-
-                // Remove aggregate-specific fields from the inner content
-                if let Some(inner_obj) = explain.as_object_mut() {
-                    for field in &Self::AGGREGATE_EXPLAIN_FIELDS {
-                        inner_obj.remove(*field);
-                    }
-                }
-            } else {
-                break;
-            }
-        }
-        explain
-    }
-
-    /// Build the explain output for a top-level aggregate query.
-    ///
-    /// Go's format: { "topLevelNode": [ {selectTopNode: ...}, {sumNode: {}}, {countNode: {}}, ... ] }
-    fn build_top_level_aggregate_explain(
-        &self,
-        select: &Select,
-        select_explain: JsonValue,
-        explain_type: ExplainType,
-    ) -> JsonValue {
-        use crate::mapper::AggregateType;
-
-        // Strip aggregate wrappers from the plan explain to get the inner selectNode content.
-        // The Rust planner wraps aggregates around the plan, but Go puts them as siblings.
-        let inner_explain = Self::strip_aggregate_wrappers(select_explain);
-
-        let mut top_level_children: Vec<JsonValue> = Vec::new();
-
-        // First element: the data source (selectTopNode)
-        top_level_children.push(serde_json::json!({
-            "selectTopNode": inner_explain
-        }));
-
-        // Add aggregate nodes based on what's in the fields
-        for field in &select.fields {
-            if let Requestable::Aggregate(agg) = field {
-                let node_name = match agg.aggregate_type {
-                    AggregateType::Sum => "sumNode",
-                    AggregateType::Count => "countNode",
-                    AggregateType::Average => "averageNode",
-                    AggregateType::Min => "minNode",
-                    AggregateType::Max => "maxNode",
-                };
-
-                // For execute explain, aggregate nodes show iterations instead of sources
-                if explain_type == ExplainType::Execute {
-                    if agg.aggregate_type == AggregateType::Average {
-                        // Go decomposes average into sumNode + countNode + averageNode
-                        // Each shows iterations: 1 in execute mode
-                        top_level_children.push(serde_json::json!({
-                            "sumNode": { "iterations": 1u64 }
-                        }));
-                        top_level_children.push(serde_json::json!({
-                            "countNode": { "iterations": 1u64 }
-                        }));
-                        top_level_children.push(serde_json::json!({
-                            "averageNode": { "iterations": 1u64 }
-                        }));
-                    } else {
-                        top_level_children.push(serde_json::json!({
-                            node_name: { "iterations": 1u64 }
-                        }));
-                    }
-                    continue;
-                }
-
-                // Default/Debug explain: show sources metadata
-                let target_filter = if !agg.targets.is_empty() {
-                    agg.targets[0].filter.as_ref()
-                } else {
-                    None
-                };
-
-                let filter_value = if let Some(filter) = target_filter {
-                    let conditions = filter.conditions();
-                    if conditions.is_empty() {
-                        JsonValue::Null
-                    } else {
-                        serde_json::json!(conditions)
-                    }
-                } else {
-                    JsonValue::Null
-                };
-
-                // For aggregates that operate on a field (sum, min, max, avg), include childFieldName
-                let child_field_name = if !agg.targets.is_empty() {
-                    agg.targets[0].field_name.as_ref()
-                } else {
-                    None
-                };
-
-                // Go decomposes average into sumNode + countNode + averageNode
-                if agg.aggregate_type == AggregateType::Average {
-                    // Go adds {field: {_neq: null}} for both sum and count source filters,
-                    // but only for regular fields (not aggregate refs like _avg).
-                    let avg_filter = if let Some(field_name) = child_field_name {
-                        if field_name.starts_with('_') {
-                            // Aggregate field refs don't get neq filter
-                            filter_value.clone()
-                        } else if filter_value.is_null() {
-                            serde_json::json!({field_name: {"_neq": serde_json::Value::Null}})
-                        } else if let Some(obj) = filter_value.as_object() {
-                            // Merge {field: {_neq: null}} into existing filter conditions
-                            let mut merged = obj.clone();
-                            merged
-                                .entry(field_name.to_string())
-                                .and_modify(|v| {
-                                    if let JsonValue::Object(ref mut ops) = v {
-                                        ops.insert("_neq".to_string(), serde_json::Value::Null);
-                                    }
-                                })
-                                .or_insert(serde_json::json!({"_neq": serde_json::Value::Null}));
-                            JsonValue::Object(merged)
-                        } else {
-                            serde_json::json!({field_name: {"_neq": serde_json::Value::Null}})
-                        }
-                    } else {
-                        filter_value.clone()
-                    };
-
-                    // 1. sumNode with sources (includes childFieldName)
-                    let sum_source = if let Some(field_name) = child_field_name {
-                        serde_json::json!({
-                            "fieldName": select.collection_name,
-                            "childFieldName": field_name,
-                            "filter": avg_filter
-                        })
-                    } else {
-                        serde_json::json!({
-                            "fieldName": select.collection_name,
-                            "filter": avg_filter
-                        })
-                    };
-                    top_level_children.push(serde_json::json!({
-                        "sumNode": {
-                            "sources": [sum_source]
-                        }
-                    }));
-
-                    // 2. countNode with sources (no childFieldName, same filter as sum)
-                    let count_source = serde_json::json!({
-                        "fieldName": select.collection_name,
-                        "filter": avg_filter
-                    });
-                    top_level_children.push(serde_json::json!({
-                        "countNode": {
-                            "sources": [count_source]
-                        }
-                    }));
-
-                    // 3. averageNode (empty)
-                    top_level_children.push(serde_json::json!({
-                        "averageNode": {}
-                    }));
-                    continue;
-                }
-
-                let source = if let Some(field_name) = child_field_name {
-                    serde_json::json!({
-                        "fieldName": select.collection_name,
-                        "childFieldName": field_name,
-                        "filter": filter_value
-                    })
-                } else {
-                    serde_json::json!({
-                        "fieldName": select.collection_name,
-                        "filter": filter_value
-                    })
-                };
-
-                top_level_children.push(serde_json::json!({
-                    node_name: {
-                        "sources": [source]
-                    }
-                }));
-            }
-        }
-
-        serde_json::json!({
-            "topLevelNode": top_level_children
-        })
     }
 
     /// Execute a GraphQL query with a specific fetcher and identity.
@@ -1931,7 +453,7 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
     /// The Planner builds a proper join plan with TypeJoinOne/TypeJoinMany nodes.
     /// ScanNodes fetch their own data via the attached fetcher.
     /// ACP permission filtering is applied per-collection via PermissionFilterNode in the plan.
-    async fn execute_nested_select_with_planner(
+    pub(crate) async fn execute_nested_select_with_planner(
         &self,
         select: &Select,
         fetcher: &dyn DocFetcher,
@@ -2688,7 +1210,7 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
     /// Execute a simple query without nested selections.
     ///
     /// This is the optimized path that supports aggregations and grouping.
-    async fn execute_simple_select(
+    pub(crate) async fn execute_simple_select(
         &self,
         select: &Select,
         fetcher: &dyn DocFetcher,
@@ -2817,699 +1339,7 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
     }
 
     /// Execute a CID-based time-travel query with optional _version support.
-    ///
-    /// Reconstructs the document as it existed at the specified commit CID
-    /// by walking the merkle DAG backwards and replaying CRDT deltas.
-    ///
-    /// CID queries require `cid` argument and optionally `docID` for validation.
-    /// For document CIDs, returns a single-element array. For collection CIDs
-    /// (branchable collections), returns all documents visible at that state.
-    async fn execute_cid_query_with_version(
-        &self,
-        select: &Select,
-        fetcher: &dyn DocFetcher,
-        _caller_identity: Option<Did>,
-        version_selection: Option<&Select>,
-    ) -> Result<JsonValue> {
-        let cid = select.cid.as_ref().ok_or_else(|| {
-            QueryError::internal("execute_cid_query called without CID - this is a bug")
-        })?;
 
-        // Get expected docID from select.doc_ids (optional validation)
-        let expected_doc_id = select.doc_ids.as_ref().and_then(|ids| ids.first());
-
-        // Fetch document(s) at the specified CID.
-        // For collection-level CIDs (branchable), this returns multiple documents.
-        // For document-level CIDs, this returns a single document.
-        let documents = match fetcher
-            .get_documents_at_cid(cid, expected_doc_id.map(|s| s.as_str()))
-            .await
-        {
-            Ok(docs) => docs,
-            Err(e) => {
-                let err_msg = e.to_string();
-                // docID mismatch: Go returns empty results
-                if err_msg.contains("cid either does not exist or belong to document") {
-                    return Ok(JsonValue::Array(vec![]));
-                }
-                // Block not found in blockstore: propagate as error (Go does the same)
-                return Err(e);
-            }
-        };
-
-        // Get collection schema for building the mapping
-        let collection = self.get_collection(&select.collection_name).await?;
-
-        // Separate nested selects (relation fields) from scalar fields.
-        // build_mapping can't handle nested selects, so we strip them and resolve relations separately.
-        let mut nested_selects: Vec<&Select> = Vec::new();
-        let scalar_fields: Vec<Requestable> = select
-            .fields
-            .iter()
-            .filter(|f| {
-                if let Requestable::Select(s) = f {
-                    if s.field.name == "_version" {
-                        return false;
-                    }
-                    nested_selects.push(s);
-                    return false;
-                }
-                true
-            })
-            .cloned()
-            .collect();
-
-        let select_for_mapping = Select {
-            fields: scalar_fields,
-            ..select.clone()
-        };
-
-        // Build mapping for scalar fields only
-        let mapping = plan::build_mapping(&select_for_mapping, &collection)?;
-
-        // Process each document into a JSON object
-        let mut result_array = Vec::new();
-
-        for document in &documents {
-            // Convert the document to JSON with only the requested scalar fields
-            let mut obj = serde_json::Map::new();
-
-            for render_key in &mapping.render_keys {
-                let field_name = mapping
-                    .try_find_name_from_index(render_key.index)
-                    .unwrap_or("");
-
-                let value = if field_name == "__typename" {
-                    JsonValue::String(select.collection_name.clone())
-                } else if field_name == "_docID" {
-                    document
-                        .id()
-                        .map(|id| JsonValue::String(id.to_string()))
-                        .unwrap_or(JsonValue::Null)
-                } else if field_name == "_deleted" {
-                    JsonValue::Bool(document.is_deleted())
-                } else if let Some(nv) = document.get(field_name) {
-                    crate::json_convert::normal_value_to_json(nv).unwrap_or(JsonValue::Null)
-                } else {
-                    JsonValue::Null
-                };
-
-                obj.insert(render_key.key.clone(), value);
-            }
-
-            // Resolve nested selects (relation fields like `author { name }`)
-            for nested_select in &nested_selects {
-                let relation_name = &nested_select.field.name;
-                let output_name = nested_select.field.output_name();
-
-                // Resolve the actual collection name from the relation field.
-                // nested_select.collection_name is the field name (e.g., "author"),
-                // but we need the collection type name (e.g., "Author").
-                let related_collection = self
-                    .resolve_relation_target_name(&collection, relation_name)
-                    .await
-                    .unwrap_or_else(|| nested_select.collection_name.clone());
-
-                // Many-to-one: parent has FK field (e.g., Book._authorID → Author)
-                let fk_field_name = CollectionVersion::relation_id_field_name(relation_name);
-                if let Some(fk_value) = document.get(&fk_field_name) {
-                    let fk_doc_id = crate::json_convert::normal_value_to_json(fk_value)
-                        .ok()
-                        .and_then(|v| v.as_str().map(|s| s.to_string()))
-                        .unwrap_or_default();
-
-                    if !fk_doc_id.is_empty() {
-                        let result = fetcher
-                            .get_by_ids(&related_collection, &[fk_doc_id])
-                            .await?;
-
-                        if let Some(related_doc) = result.docs().first() {
-                            let related_obj =
-                                self.render_document_fields(related_doc, nested_select);
-                            obj.insert(output_name.to_string(), JsonValue::Object(related_obj));
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Null);
-                        }
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Null);
-                    }
-                } else {
-                    // One-to-many or no FK found: return null for now
-                    obj.insert(output_name.to_string(), JsonValue::Null);
-                }
-            }
-
-            // Add _version data if requested
-            if let Some(version_select) = version_selection {
-                let doc_id = document.id().map(|id| id.to_string());
-                if let Some(doc_id_str) = doc_id {
-                    let version_data = self
-                        .fetch_version_data(fetcher, &doc_id_str, version_select, Some(cid))
-                        .await?;
-                    let output_name = version_select.field.output_name();
-                    obj.insert(output_name.to_string(), version_data);
-                }
-            }
-
-            result_array.push(JsonValue::Object(obj));
-        }
-
-        Ok(JsonValue::Array(result_array))
-    }
-
-    /// Execute a regular query with _version field support.
-    ///
-    /// This handles queries that include _version selection but don't have a CID argument.
-    /// For each document result, fetches the commit history and adds _version data.
-    async fn execute_query_with_version(
-        &self,
-        select: &Select,
-        fetcher: &dyn DocFetcher,
-        caller_identity: Option<Did>,
-        version_selection: Option<&Select>,
-    ) -> Result<JsonValue> {
-        // Get collection schema
-        let collection = self.get_collection(&select.collection_name).await?;
-
-        // Check if _docID is already in the selection (we need it to fetch version data)
-        let has_doc_id = select.fields.iter().any(|f| {
-            if let Requestable::Field(field) = f {
-                field.name == "_docID"
-            } else {
-                false
-            }
-        });
-
-        // Build a modified select without _version for the regular query
-        // (We'll add _version data after fetching documents)
-        // Also add _docID if not already present (needed to fetch version data)
-        let mut fields_without_version: Vec<Requestable> = select
-            .fields
-            .iter()
-            .filter(|f| {
-                if let Requestable::Select(s) = f {
-                    s.field.name != "_version"
-                } else {
-                    true
-                }
-            })
-            .cloned()
-            .collect();
-
-        // Add _docID field if not already present
-        if !has_doc_id {
-            fields_without_version.push(Requestable::Field(crate::mapper::Field {
-                name: "_docID".to_string(),
-                alias: None,
-            }));
-        }
-
-        let select_without_version = Select {
-            fields: fields_without_version,
-            ..select.clone()
-        };
-
-        // Check if remaining fields need the Planner (has real nested selections)
-        let has_nested = select_without_version
-            .fields
-            .iter()
-            .any(|f| matches!(f, Requestable::Select(_)));
-
-        let filter_has_relations = select
-            .filter
-            .as_ref()
-            .map(|f| f.has_relation_filters())
-            .unwrap_or(false);
-
-        let order_has_relations = select
-            .order_by
-            .as_ref()
-            .map(|o| o.has_relation_order())
-            .unwrap_or(false);
-
-        // Views always need the planner (they execute queries, not storage reads)
-        let is_view = collection.query.is_some();
-
-        // Execute the query for document data (without _version)
-        let result = if is_view || has_nested || filter_has_relations || order_has_relations {
-            self.execute_nested_select_with_planner(
-                &select_without_version,
-                fetcher,
-                caller_identity,
-            )
-            .await?
-        } else {
-            self.execute_simple_select(
-                &select_without_version,
-                fetcher,
-                &collection,
-                caller_identity,
-            )
-            .await?
-        };
-
-        // If no _version selection, return as-is
-        let version_select = match version_selection {
-            Some(v) => v,
-            None => return Ok(result),
-        };
-
-        // Add _version data to each document result
-        let results = result
-            .as_array()
-            .ok_or_else(|| QueryError::internal("Expected array result"))?;
-
-        let mut enriched_results = Vec::new();
-        for doc_json in results {
-            let mut doc_obj = doc_json
-                .as_object()
-                .ok_or_else(|| QueryError::internal("Expected object in result"))?
-                .clone();
-
-            // Get document ID from the result
-            let doc_id = doc_obj
-                .get("_docID")
-                .and_then(|v| v.as_str())
-                .map(|s| s.to_string());
-
-            // Remove _docID if it wasn't originally requested
-            if !has_doc_id {
-                doc_obj.remove("_docID");
-            }
-
-            if let Some(doc_id_str) = doc_id {
-                let version_data = self
-                    .fetch_version_data(fetcher, &doc_id_str, version_select, None)
-                    .await?;
-                let output_name = version_select.field.output_name();
-                doc_obj.insert(output_name.to_string(), version_data);
-            } else {
-                // No docID available - return empty version array
-                let output_name = version_select.field.output_name();
-                doc_obj.insert(output_name.to_string(), JsonValue::Array(vec![]));
-            }
-
-            enriched_results.push(JsonValue::Object(doc_obj));
-        }
-
-        Ok(JsonValue::Array(enriched_results))
-    }
-
-    /// Resolve the actual collection name for a relation field on a parent collection.
-    ///
-    /// Nested selects have collection_name set to the field name (e.g., "author"),
-    /// but fetcher operations need the collection type name (e.g., "Author").
-    /// This method resolves the target collection name by looking at the parent
-    /// collection's relation field metadata.
-    async fn resolve_relation_target_name(
-        &self,
-        parent_collection: &CollectionVersion,
-        relation_field_name: &str,
-    ) -> Option<String> {
-        let field = parent_collection.field_by_name(relation_field_name)?;
-        let target_id = field.kind.relation_collection_id()?;
-
-        // For Named fields, target_id is the collection name directly
-        if let Ok(Some(coll)) = self.collection_provider.get_collection(target_id).await {
-            return Some(coll.name.clone());
-        }
-
-        // For Relation fields, target_id is a collection_id/version_id — search by ID
-        if let Ok(names) = self.collection_provider.list_collections().await {
-            for name in names {
-                if let Ok(Some(coll)) = self.collection_provider.get_collection(&name).await {
-                    if coll.collection_id == target_id || coll.version_id == target_id {
-                        return Some(coll.name.clone());
-                    }
-                }
-            }
-        }
-        None
-    }
-
-    /// Render a Document's fields as a JSON object using only the fields requested by a Select.
-    fn render_document_fields(
-        &self,
-        doc: &Document,
-        select: &Select,
-    ) -> serde_json::Map<String, JsonValue> {
-        let mut obj = serde_json::Map::new();
-        for field in &select.fields {
-            if let Requestable::Field(f) = field {
-                let fname = &f.name;
-                let output = f.output_name();
-                if fname == "_docID" {
-                    if let Some(id) = doc.id() {
-                        obj.insert(output.to_string(), JsonValue::String(id.to_string()));
-                    } else {
-                        obj.insert(output.to_string(), JsonValue::Null);
-                    }
-                } else if fname == "__typename" {
-                    obj.insert(
-                        output.to_string(),
-                        JsonValue::String(select.collection_name.clone()),
-                    );
-                } else if let Some(nv) = doc.get(fname) {
-                    let json_val =
-                        crate::json_convert::normal_value_to_json(nv).unwrap_or(JsonValue::Null);
-                    obj.insert(output.to_string(), json_val);
-                } else {
-                    obj.insert(output.to_string(), JsonValue::Null);
-                }
-            }
-        }
-        obj
-    }
-
-    /// Fetch version (commit) data for a document.
-    ///
-    /// Returns an array of commit objects filtered to composite commits (fieldName = "_C")
-    /// and rendered with the requested fields from the _version selection.
-    pub(crate) async fn fetch_version_data(
-        &self,
-        fetcher: &dyn DocFetcher,
-        doc_id: &str,
-        version_select: &Select,
-        target_cid: Option<&str>,
-    ) -> Result<JsonValue> {
-        use crate::fetcher::CommitsQueryOptions;
-
-        // Fetch commits for this document
-        // When we have a target CID, we need to traverse all commits back to genesis
-        // by setting a large depth. Without a target CID (regular query), depth=None
-        // traverses all heads to genesis anyway.
-        let depth = if target_cid.is_some() {
-            Some(1000) // Reasonable max depth for version history traversal
-        } else {
-            None
-        };
-
-        let options = CommitsQueryOptions {
-            doc_id: Some(doc_id.to_string()),
-            cid: target_cid.map(|s| s.to_string()),
-            depth,
-            field_name: None,
-        };
-
-        let commits = fetcher.get_commits(&options).await?;
-
-        // Filter to composite commits only (fieldName = "_C")
-        // and render the requested fields
-        let mut version_results: Vec<JsonValue> = Vec::new();
-
-        for commit in commits {
-            // Filter to composite commits
-            let field_name = commit
-                .get("fieldName")
-                .and_then(|v| v.as_str())
-                .unwrap_or("");
-            if field_name != "_C" {
-                continue;
-            }
-
-            let commit_json = self.render_commit(&commit, version_select)?;
-            version_results.push(commit_json);
-        }
-
-        // Sort by height descending (newest first)
-        version_results.sort_by(|a, b| {
-            let h_a = a.get("height").and_then(|v| v.as_i64()).unwrap_or(0);
-            let h_b = b.get("height").and_then(|v| v.as_i64()).unwrap_or(0);
-            h_b.cmp(&h_a)
-        });
-
-        Ok(JsonValue::Array(version_results))
-    }
-
-    /// Render a commit document according to the _version selection fields.
-    fn render_commit(&self, commit: &Document, version_select: &Select) -> Result<JsonValue> {
-        let mut obj = serde_json::Map::new();
-
-        for requestable in &version_select.fields {
-            match requestable {
-                Requestable::Field(f) => {
-                    let field_name = &f.name;
-                    let output_name = f.output_name();
-
-                    if let Some(value) = commit.get(field_name) {
-                        let json_value = crate::json_convert::normal_value_to_json(value)
-                            .unwrap_or(JsonValue::Null);
-                        obj.insert(output_name.to_string(), json_value);
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Null);
-                    }
-                }
-                Requestable::Select(nested) => {
-                    let field_name = &nested.field.name;
-                    let output_name = nested.field.output_name();
-
-                    // Handle nested selections (links, heads) with optional filter
-                    if let Some(value) = commit.get(field_name) {
-                        if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
-                            if let Some(arr) = json_val.as_array() {
-                                // Apply filter if present on the nested selection
-                                let filtered_items: Vec<&JsonValue> =
-                                    if let Some(ref filter) = nested.filter {
-                                        arr.iter()
-                                            .filter(|item| {
-                                                // Check each filter condition against the item
-                                                self.json_item_matches_filter(item, filter)
-                                            })
-                                            .collect()
-                                    } else {
-                                        arr.iter().collect()
-                                    };
-
-                                let nested_results: Vec<JsonValue> = filtered_items
-                                    .into_iter()
-                                    .map(|item| {
-                                        let mut nested_obj = serde_json::Map::new();
-                                        for nested_field in &nested.fields {
-                                            if let Requestable::Field(nf) = nested_field {
-                                                let nf_name = &nf.name;
-                                                let nf_output = nf.output_name();
-                                                if let Some(nv) = item.get(nf_name) {
-                                                    nested_obj
-                                                        .insert(nf_output.to_string(), nv.clone());
-                                                } else {
-                                                    nested_obj.insert(
-                                                        nf_output.to_string(),
-                                                        JsonValue::Null,
-                                                    );
-                                                }
-                                            }
-                                        }
-                                        JsonValue::Object(nested_obj)
-                                    })
-                                    .collect();
-                                obj.insert(
-                                    output_name.to_string(),
-                                    JsonValue::Array(nested_results),
-                                );
-                            } else {
-                                obj.insert(output_name.to_string(), JsonValue::Null);
-                            }
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Null);
-                        }
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Array(vec![]));
-                    }
-                }
-                Requestable::Aggregate(agg) => {
-                    // Handle aggregates on commit fields (e.g., _count(links: {}))
-                    let output_name = agg.output_name();
-                    if let Some(target) = agg.targets.first() {
-                        let target_field = target
-                            .field_name
-                            .as_deref()
-                            .filter(|s| !s.is_empty())
-                            .or_else(|| Some(target.host_name.as_str()).filter(|s| !s.is_empty()));
-
-                        if let Some(field) = target_field {
-                            if let Some(val) = commit.get(field) {
-                                if let Ok(json_val) = crate::json_convert::normal_value_to_json(val)
-                                {
-                                    if let Some(arr) = json_val.as_array() {
-                                        obj.insert(
-                                            output_name.to_string(),
-                                            JsonValue::Number((arr.len() as i64).into()),
-                                        );
-                                    } else {
-                                        obj.insert(
-                                            output_name.to_string(),
-                                            JsonValue::Number(0.into()),
-                                        );
-                                    }
-                                } else {
-                                    obj.insert(
-                                        output_name.to_string(),
-                                        JsonValue::Number(0.into()),
-                                    );
-                                }
-                            } else {
-                                obj.insert(output_name.to_string(), JsonValue::Number(0.into()));
-                            }
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Number(1.into()));
-                        }
-                    } else {
-                        obj.insert(output_name.to_string(), JsonValue::Number(1.into()));
-                    }
-                }
-                Requestable::Similarity(_) => {
-                    // Similarity is not applicable in commit context
-                }
-            }
-        }
-
-        Ok(JsonValue::Object(obj))
-    }
-
-    /// Check if a JSON object matches a filter for nested commit selections.
-    ///
-    /// This is a simplified filter matcher for nested selections like `links(filter: {fieldName: {_eq: "Age"}})`.
-    /// The filter conditions are stored as `{field_name: {_op: value}}`.
-    fn json_item_matches_filter(&self, item: &JsonValue, filter: &crate::mapper::Filter) -> bool {
-        use crate::mapper::FilterOp;
-
-        // Get the filter conditions - a map of field_name -> operator conditions
-        let conditions = filter.conditions();
-
-        for (field_name, condition_value) in conditions {
-            // Check if this is a logical operator (_and, _or, _not)
-            if let Some(op) = FilterOp::parse(field_name) {
-                match op {
-                    FilterOp::And => {
-                        if let JsonValue::Array(arr) = condition_value {
-                            for sub_cond in arr {
-                                if let JsonValue::Object(obj) = sub_cond {
-                                    let sub_map: std::collections::HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    let sub_filter =
-                                        crate::mapper::Filter::from_conditions(sub_map);
-                                    if !self.json_item_matches_filter(item, &sub_filter) {
-                                        return false;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FilterOp::Or => {
-                        if let JsonValue::Array(arr) = condition_value {
-                            let mut any_match = false;
-                            for sub_cond in arr {
-                                if let JsonValue::Object(obj) = sub_cond {
-                                    let sub_map: std::collections::HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    let sub_filter =
-                                        crate::mapper::Filter::from_conditions(sub_map);
-                                    if self.json_item_matches_filter(item, &sub_filter) {
-                                        any_match = true;
-                                        break;
-                                    }
-                                }
-                            }
-                            if !any_match {
-                                return false;
-                            }
-                        }
-                    }
-                    FilterOp::Not => {
-                        if let JsonValue::Object(obj) = condition_value {
-                            let sub_map: std::collections::HashMap<String, JsonValue> =
-                                obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                            let sub_filter = crate::mapper::Filter::from_conditions(sub_map);
-                            if self.json_item_matches_filter(item, &sub_filter) {
-                                return false;
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-                continue;
-            }
-
-            // This is a field condition: field_name -> {_op: value}
-            let item_value = item.get(field_name);
-
-            // The condition_value should be an object like {"_eq": "Age"}
-            if let JsonValue::Object(ops) = condition_value {
-                for (op_name, expected_value) in ops {
-                    if let Some(op) = FilterOp::parse(op_name) {
-                        let matches = self.check_filter_op(item_value, op, expected_value);
-                        if !matches {
-                            return false;
-                        }
-                    }
-                }
-            }
-        }
-
-        true
-    }
-
-    /// Check if an item value matches a filter operator condition.
-    fn check_filter_op(
-        &self,
-        item_value: Option<&JsonValue>,
-        op: crate::mapper::FilterOp,
-        expected: &JsonValue,
-    ) -> bool {
-        use crate::mapper::FilterOp;
-
-        match op {
-            FilterOp::Eq => match (item_value, expected) {
-                (Some(JsonValue::String(a)), JsonValue::String(b)) => a == b,
-                (Some(JsonValue::Number(a)), JsonValue::Number(b)) => a == b,
-                (Some(JsonValue::Bool(a)), JsonValue::Bool(b)) => a == b,
-                (Some(JsonValue::Null), JsonValue::Null) => true,
-                (None, JsonValue::Null) => true,
-                _ => false,
-            },
-            FilterOp::Ne => match (item_value, expected) {
-                (Some(JsonValue::String(a)), JsonValue::String(b)) => a != b,
-                (Some(JsonValue::Number(a)), JsonValue::Number(b)) => a != b,
-                (Some(JsonValue::Bool(a)), JsonValue::Bool(b)) => a != b,
-                (Some(JsonValue::Null), JsonValue::Null) => false,
-                (None, _) => true,
-                _ => true,
-            },
-            FilterOp::Gt => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a > b,
-                _ => false,
-            },
-            FilterOp::Gte => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a >= b,
-                _ => false,
-            },
-            FilterOp::Lt => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a < b,
-                _ => false,
-            },
-            FilterOp::Lte => match (item_value.and_then(|v| v.as_f64()), expected.as_f64()) {
-                (Some(a), Some(b)) => a <= b,
-                _ => false,
-            },
-            FilterOp::In => {
-                if let JsonValue::Array(values) = expected {
-                    item_value.map(|v| values.contains(v)).unwrap_or(false)
-                } else {
-                    false
-                }
-            }
-            FilterOp::Nin => {
-                if let JsonValue::Array(values) = expected {
-                    item_value.map(|v| !values.contains(v)).unwrap_or(true)
-                } else {
-                    true
-                }
-            }
-            _ => true, // For unsupported operators, default to matching
-        }
-    }
-
-    /// Execute a top-level aggregate query (e.g., `{ _avg(Users: {field: Age}) }`).
     ///
     /// Top-level aggregates compute a single value over all documents in a collection.
     /// Unlike regular collection queries that return an array, top-level aggregates
@@ -3882,513 +1712,6 @@ impl<F: DocFetcher + 'static, R: TransactionRegistry> QueryRunner<F, R> {
         // The caller will insert this with the correct key
         let _ = output_name; // suppress unused warning
         Ok(value)
-    }
-
-    /// Execute a _commits system collection query.
-    ///
-    /// This handles queries to the special _commits collection which fetches
-    /// commit history from the headstore and blockstore.
-    async fn execute_commits_query(&self, select: &Select) -> Result<JsonValue> {
-        use crate::fetcher::CommitsQueryOptions;
-        use crate::mapper::{AggregateType, OrderDirection};
-
-        // Build options from the select
-        let options = CommitsQueryOptions {
-            doc_id: select.doc_ids.as_ref().and_then(|ids| ids.first().cloned()),
-            cid: select.cid.clone(),
-            depth: select.depth,
-            field_name: None,
-        };
-
-        // Fetch commits using the fetcher
-        let mut commits = self.fetcher.get_commits(&options).await?;
-
-        // Build a mapping for commit fields (needed for filter evaluation)
-        let mapping = Self::build_commits_mapping();
-
-        // Apply filter if present
-        if let Some(ref filter) = select.filter {
-            commits.retain(|commit| {
-                let fields = Self::commit_to_fields(commit, &mapping);
-                filter.matches(&fields, &mapping).unwrap_or(true)
-            });
-        }
-
-        // Apply groupBy if present - this changes how we process commits
-        // Each entry is (representative_commit, all_commits_in_group)
-        let grouped: Option<Vec<(document::Document, Vec<document::Document>)>> =
-            if let Some(ref group_by) = select.group_by {
-                if !group_by.fields.is_empty() {
-                    let mut groups: Vec<(String, document::Document, Vec<document::Document>)> =
-                        Vec::new();
-                    let mut group_map: std::collections::HashMap<String, usize> =
-                        std::collections::HashMap::new();
-
-                    for commit in commits.drain(..) {
-                        let key = Self::generate_commit_group_key(&commit, &group_by.fields);
-                        if let Some(&idx) = group_map.get(&key) {
-                            groups[idx].2.push(commit);
-                        } else {
-                            let idx = groups.len();
-                            group_map.insert(key.clone(), idx);
-                            groups.push((key, commit.clone(), vec![commit]));
-                        }
-                    }
-
-                    Some(
-                        groups
-                            .into_iter()
-                            .map(|(_, rep, docs)| (rep, docs))
-                            .collect(),
-                    )
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-        // Get the list of commits to process (either grouped representatives or all commits)
-        let mut work_items: Vec<(document::Document, Option<Vec<document::Document>>)> =
-            if let Some(grouped) = grouped {
-                grouped
-                    .into_iter()
-                    .map(|(rep, all)| (rep, Some(all)))
-                    .collect()
-            } else {
-                commits.into_iter().map(|c| (c, None)).collect()
-            };
-
-        // Apply ordering if present
-        if let Some(ref order_by) = select.order_by {
-            for condition in order_by.conditions.iter().rev() {
-                if let Some(field_name) = condition.fields.first() {
-                    let desc = matches!(condition.direction, OrderDirection::Desc);
-                    work_items.sort_by(|(a, _), (b, _)| {
-                        let val_a = a.get(field_name);
-                        let val_b = b.get(field_name);
-                        let cmp = Self::compare_json_values(val_a, val_b);
-                        if desc {
-                            cmp.reverse()
-                        } else {
-                            cmp
-                        }
-                    });
-                }
-            }
-        }
-
-        // Apply limit and offset if present
-        if let Some(ref limit_spec) = select.limit {
-            let offset = limit_spec.offset as usize;
-            if offset > 0 && offset < work_items.len() {
-                work_items = work_items.split_off(offset);
-            } else if offset >= work_items.len() {
-                work_items.clear();
-            }
-            if let Some(limit) = limit_spec.limit {
-                work_items.truncate(limit as usize);
-            }
-        }
-
-        // Build results
-        let mut results = Vec::new();
-        for (commit, group_docs) in &work_items {
-            let mut obj = serde_json::Map::new();
-
-            // Map requested fields from the commit document
-            for field in &select.fields {
-                match field {
-                    Requestable::Field(f) => {
-                        let field_name = &f.name;
-                        let output_name = f.output_name();
-
-                        // Handle __typename specially for commits (Go returns "Commit")
-                        if field_name == "__typename" {
-                            obj.insert(
-                                output_name.to_string(),
-                                JsonValue::String("Commit".to_string()),
-                            );
-                        } else if let Some(value) = commit.get(field_name) {
-                            let json_value = crate::json_convert::normal_value_to_json(value)
-                                .unwrap_or(JsonValue::Null);
-                            obj.insert(output_name.to_string(), json_value);
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Null);
-                        }
-                    }
-                    Requestable::Aggregate(agg) => {
-                        // Handle aggregates like _count on links/heads
-                        let output_name = agg.output_name();
-                        let count = match agg.aggregate_type {
-                            AggregateType::Count => {
-                                // Count on a target field (e.g., links, heads)
-                                if let Some(target) = agg.targets.first() {
-                                    // Get the target field name - check field_name first (from
-                                    // `field:` arg), then host_name (from relation syntax)
-                                    let target_field = target
-                                        .field_name
-                                        .as_deref()
-                                        .filter(|s| !s.is_empty())
-                                        .or_else(|| {
-                                            Some(target.host_name.as_str())
-                                                .filter(|s| !s.is_empty())
-                                        });
-
-                                    if let Some(field) = target_field {
-                                        if let Some(val) = commit.get(field) {
-                                            // Convert NormalValue to JSON to check array
-                                            if let Ok(json_val) =
-                                                crate::json_convert::normal_value_to_json(val)
-                                            {
-                                                if let Some(arr) = json_val.as_array() {
-                                                    arr.len() as i64
-                                                } else {
-                                                    0
-                                                }
-                                            } else {
-                                                0
-                                            }
-                                        } else {
-                                            0
-                                        }
-                                    } else {
-                                        1 // Count without target = count this commit
-                                    }
-                                } else {
-                                    1 // Count without target = count this commit
-                                }
-                            }
-                            _ => 0, // Other aggregates not supported for commits
-                        };
-                        obj.insert(output_name.to_string(), JsonValue::Number(count.into()));
-                    }
-                    Requestable::Select(nested) => {
-                        let field_name = &nested.field.name;
-                        let output_name = nested.field.output_name();
-
-                        // Handle _group special field for grouped results
-                        if field_name == "_group" {
-                            if let Some(docs) = group_docs {
-                                // Build array of group documents with requested fields
-                                let group_array: Vec<JsonValue> = docs
-                                    .iter()
-                                    .map(|doc: &document::Document| {
-                                        let mut nested_obj = serde_json::Map::new();
-                                        for nested_field in &nested.fields {
-                                            if let Requestable::Field(nf) = nested_field {
-                                                let nf_name = &nf.name;
-                                                let nf_output = nf.output_name();
-                                                if let Some(val) = doc.get(nf_name) {
-                                                    let json_val =
-                                                        crate::json_convert::normal_value_to_json(
-                                                            val,
-                                                        )
-                                                        .unwrap_or(JsonValue::Null);
-                                                    nested_obj
-                                                        .insert(nf_output.to_string(), json_val);
-                                                } else {
-                                                    nested_obj.insert(
-                                                        nf_output.to_string(),
-                                                        JsonValue::Null,
-                                                    );
-                                                }
-                                            }
-                                        }
-                                        JsonValue::Object(nested_obj)
-                                    })
-                                    .collect();
-                                obj.insert(output_name.to_string(), JsonValue::Array(group_array));
-                            } else {
-                                // Not grouped, _group is empty
-                                obj.insert(output_name.to_string(), JsonValue::Array(vec![]));
-                            }
-                        } else if let Some(value) = commit.get(field_name) {
-                            // Handle nested selects (e.g., links { cid })
-                            if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
-                                if let Some(arr) = json_val.as_array() {
-                                    // Handle array nested selects (e.g., links { cid }, heads { cid })
-                                    // Apply filter if present (e.g., links(filter: {fieldName: {_eq: "age"}}))
-                                    let nested_results: Vec<JsonValue> = arr
-                                        .iter()
-                                        .filter(|item| {
-                                            // Apply nested filter if present
-                                            if let Some(ref filter) = nested.filter {
-                                                Self::matches_json_filter(item, filter)
-                                            } else {
-                                                true
-                                            }
-                                        })
-                                        .map(|item: &JsonValue| {
-                                            let mut nested_obj = serde_json::Map::new();
-                                            for nested_field in &nested.fields {
-                                                if let Requestable::Field(nf) = nested_field {
-                                                    let nf_name = &nf.name;
-                                                    let nf_output = nf.output_name();
-                                                    if let Some(nv) = item.get(nf_name) {
-                                                        nested_obj.insert(
-                                                            nf_output.to_string(),
-                                                            nv.clone(),
-                                                        );
-                                                    } else {
-                                                        nested_obj.insert(
-                                                            nf_output.to_string(),
-                                                            JsonValue::Null,
-                                                        );
-                                                    }
-                                                } else if let Requestable::Select(inner_nested) =
-                                                    nested_field
-                                                {
-                                                    // Handle double-nested selects (e.g., heads { links { cid } })
-                                                    let inner_name = &inner_nested.field.name;
-                                                    let inner_output =
-                                                        inner_nested.field.output_name();
-                                                    if let Some(inner_val) = item.get(inner_name) {
-                                                        if let Some(inner_arr) = inner_val.as_array()
-                                                        {
-                                                            // Apply filter and extract requested fields
-                                                            let inner_results: Vec<JsonValue> =
-                                                                inner_arr
-                                                                    .iter()
-                                                                    .filter(|inner_item| {
-                                                                        if let Some(ref filter) =
-                                                                            inner_nested.filter
-                                                                        {
-                                                                            Self::matches_json_filter(
-                                                                                inner_item, filter,
-                                                                            )
-                                                                        } else {
-                                                                            true
-                                                                        }
-                                                                    })
-                                                                    .map(|inner_item| {
-                                                                        let mut inner_obj =
-                                                                            serde_json::Map::new();
-                                                                        for inner_field in
-                                                                            &inner_nested.fields
-                                                                        {
-                                                                            if let Requestable::Field(
-                                                                                inf,
-                                                                            ) = inner_field
-                                                                            {
-                                                                                let inf_name =
-                                                                                    &inf.name;
-                                                                                let inf_output =
-                                                                                    inf.output_name(
-                                                                                    );
-                                                                                if let Some(inv) =
-                                                                                    inner_item
-                                                                                        .get(inf_name)
-                                                                                {
-                                                                                    inner_obj.insert(
-                                                                                        inf_output
-                                                                                            .to_string(
-                                                                                            ),
-                                                                                        inv.clone(),
-                                                                                    );
-                                                                                } else {
-                                                                                    inner_obj.insert(
-                                                                                        inf_output
-                                                                                            .to_string(
-                                                                                            ),
-                                                                                        JsonValue::Null,
-                                                                                    );
-                                                                                }
-                                                                            }
-                                                                        }
-                                                                        JsonValue::Object(inner_obj)
-                                                                    })
-                                                                    .collect();
-                                                            nested_obj.insert(
-                                                                inner_output.to_string(),
-                                                                JsonValue::Array(inner_results),
-                                                            );
-                                                        } else {
-                                                            nested_obj.insert(
-                                                                inner_output.to_string(),
-                                                                inner_val.clone(),
-                                                            );
-                                                        }
-                                                    } else {
-                                                        nested_obj.insert(
-                                                            inner_output.to_string(),
-                                                            JsonValue::Array(vec![]),
-                                                        );
-                                                    }
-                                                }
-                                            }
-                                            JsonValue::Object(nested_obj)
-                                        })
-                                        .collect();
-                                    obj.insert(
-                                        output_name.to_string(),
-                                        JsonValue::Array(nested_results),
-                                    );
-                                } else if json_val.is_object() {
-                                    // Handle object nested selects (e.g., signature { type identity value })
-                                    let mut nested_obj = serde_json::Map::new();
-                                    for nested_field in &nested.fields {
-                                        if let Requestable::Field(nf) = nested_field {
-                                            let nf_name = &nf.name;
-                                            let nf_output = nf.output_name();
-                                            if let Some(nv) = json_val.get(nf_name) {
-                                                nested_obj
-                                                    .insert(nf_output.to_string(), nv.clone());
-                                            } else {
-                                                nested_obj
-                                                    .insert(nf_output.to_string(), JsonValue::Null);
-                                            }
-                                        }
-                                    }
-                                    obj.insert(
-                                        output_name.to_string(),
-                                        JsonValue::Object(nested_obj),
-                                    );
-                                } else {
-                                    obj.insert(output_name.to_string(), JsonValue::Null);
-                                }
-                            } else {
-                                obj.insert(output_name.to_string(), JsonValue::Null);
-                            }
-                        } else {
-                            obj.insert(output_name.to_string(), JsonValue::Array(vec![]));
-                        }
-                    }
-                    Requestable::Similarity(_) => {
-                        // Similarity is not applicable in commit context
-                    }
-                }
-            }
-
-            results.push(JsonValue::Object(obj));
-        }
-
-        Ok(JsonValue::Array(results))
-    }
-
-    /// Generate a group key from commit field values.
-    /// Format matches Go DefraDB: `{index}_{value}_` for each field.
-    fn generate_commit_group_key(commit: &document::Document, fields: &[String]) -> String {
-        let mut key = String::new();
-        for (i, field_name) in fields.iter().enumerate() {
-            key.push_str(&format!("{}_", i));
-            if let Some(value) = commit.get(field_name) {
-                if let Ok(json_val) = crate::json_convert::normal_value_to_json(value) {
-                    key.push_str(&Self::json_value_to_key(&json_val));
-                } else {
-                    key.push_str("null");
-                }
-            } else {
-                key.push_str("null");
-            }
-            key.push('_');
-        }
-        key
-    }
-
-    /// Convert a JSON value to a string for use in group key.
-    fn json_value_to_key(value: &JsonValue) -> String {
-        match value {
-            JsonValue::Null => "null".to_string(),
-            JsonValue::Bool(b) => b.to_string(),
-            JsonValue::Number(n) => n.to_string(),
-            JsonValue::String(s) => s.clone(),
-            JsonValue::Array(arr) => {
-                format!(
-                    "[{}]",
-                    arr.iter()
-                        .map(Self::json_value_to_key)
-                        .collect::<Vec<_>>()
-                        .join(",")
-                )
-            }
-            JsonValue::Object(obj) => {
-                format!(
-                    "{{{}}}",
-                    obj.iter()
-                        .map(|(k, v)| format!("{}:{}", k, Self::json_value_to_key(v)))
-                        .collect::<Vec<_>>()
-                        .join(",")
-                )
-            }
-        }
-    }
-
-    /// Check if a JSON value matches a filter.
-    /// Used for filtering nested arrays like links and heads in commit queries.
-    fn matches_json_filter(value: &JsonValue, filter: &crate::Filter) -> bool {
-        filter.matches_json_object(value).unwrap_or(false)
-    }
-
-    /// Build a DocumentMapping for commit fields.
-    fn build_commits_mapping() -> crate::document::DocumentMapping {
-        let mut mapping = crate::document::DocumentMapping::new();
-        mapping.add(0, "cid");
-        mapping.add(1, "height");
-        mapping.add(2, "fieldName");
-        mapping.add(3, "docID");
-        mapping.add(4, "delta");
-        mapping.add(5, "collectionVersionId");
-        mapping.add(6, "links");
-        mapping.add(7, "heads");
-        mapping.add(8, "signature");
-        mapping
-    }
-
-    /// Convert a commit document to a fields array for filter evaluation.
-    fn commit_to_fields(
-        commit: &document::Document,
-        _mapping: &crate::document::DocumentMapping,
-    ) -> Vec<Option<JsonValue>> {
-        let field_names = [
-            "cid",
-            "height",
-            "fieldName",
-            "docID",
-            "delta",
-            "collectionVersionId",
-            "links",
-            "heads",
-            "signature",
-        ];
-        field_names
-            .iter()
-            .map(|name| {
-                commit
-                    .get(name)
-                    .and_then(|v| crate::json_convert::normal_value_to_json(v).ok())
-            })
-            .collect()
-    }
-
-    /// Compare two JSON values for ordering.
-    fn compare_json_values(
-        a: Option<&document::NormalValue>,
-        b: Option<&document::NormalValue>,
-    ) -> std::cmp::Ordering {
-        use std::cmp::Ordering;
-
-        match (a, b) {
-            (None, None) => Ordering::Equal,
-            (None, Some(_)) => Ordering::Less,
-            (Some(_), None) => Ordering::Greater,
-            (Some(va), Some(vb)) => {
-                // Convert to JSON for comparison
-                let ja = crate::json_convert::normal_value_to_json(va).ok();
-                let jb = crate::json_convert::normal_value_to_json(vb).ok();
-
-                match (ja, jb) {
-                    (Some(JsonValue::Number(na)), Some(JsonValue::Number(nb))) => {
-                        let fa = na.as_f64().unwrap_or(0.0);
-                        let fb = nb.as_f64().unwrap_or(0.0);
-                        fa.partial_cmp(&fb).unwrap_or(Ordering::Equal)
-                    }
-                    (Some(JsonValue::String(sa)), Some(JsonValue::String(sb))) => sa.cmp(&sb),
-                    (Some(a), Some(b)) => a.to_string().cmp(&b.to_string()),
-                    _ => Ordering::Equal,
-                }
-            }
-        }
     }
 }
